+++ lib/components/encoder/encoder.h
@@ -12,16 +12,16 @@ class Encoder {
   // ISR called upon v1/encA change
   void interruptHandler();
   // Return current encoder position
-  uint8_t getPosition();
+  int16_t getPosition();
   // Define encoder position
-  void setPosition(uint8_t position);
+  void setPosition(int16_t position);
   // Update internal variables/state machine
   void schedule();
 
  private:
   hardwareAbstraction::HalInterface *_hal;
   uint8_t _pin_v1, _pin_v2;
-  uint8_t _position, _isr_position;
+  int16_t _position, _isr_position;
   bool _isr_doorbell;
 };
